Move the loss helpers out of tfjs-backend, fold binaryCrossentropy in
