+++ src/backend/tfjs_backend_test.ts
@@ -1308,97 +1308,6 @@ describe('getUID ', () => {
   });
 });
 
-describeMathCPUAndGPU('categoricalCrossentropy ', () => {
-  it('from logits', () => {
-    const x = tensor2d([[1, 2], [3, 4]], [2, 2]);
-    const target = tensor2d([[0.25, 0.75], [0.1, 0.9]], [2, 2]);
-    const expected = tensor1d([
-      -1 *
-      (Math.log(Math.exp(1) / (Math.exp(1) + Math.exp(2))) * 0.25 +
-        Math.log(Math.exp(2) / (Math.exp(1) + Math.exp(2))) * 0.75),
-      -1 *
-      (Math.log(Math.exp(3) / (Math.exp(3) + Math.exp(4))) * 0.1 +
-        Math.log(Math.exp(4) / (Math.exp(3) + Math.exp(4))) * 0.9)
-    ]);
-    const result = K.categoricalCrossentropy(target, x, true);
-    expectTensorsClose(result, expected);
-  });
-
-  it('from softmax', () => {
-    const x = tensor2d([[0.3, 0.7], [0.4, 0.6]], [2, 2]);
-    const target = tensor2d([[0.25, 0.75], [0.1, 0.9]], [2, 2]);
-    const expected = tensor1d([
-      -1 * (Math.log(0.3) * 0.25 + Math.log(0.7) * 0.75),
-      -1 * (Math.log(0.4) * 0.1 + Math.log(0.6) * 0.9)
-    ]);
-    const result = K.categoricalCrossentropy(target, x, false);
-    expectTensorsClose(result, expected);
-  });
-});
-
-describeMathCPUAndGPU('sparseCategoricalCrossentropy ', () => {
-  it('from logits', () => {
-    const x = tensor2d([[1, 2, 3], [4, 5, 6]], [2, 3]);
-    const target = tensor1d([0, 2]);
-    const expected = tensor1d([
-      -1 * Math.log(Math.exp(1) / (Math.exp(1) + Math.exp(2) + Math.exp(3))),
-      -1 * Math.log(Math.exp(6) / (Math.exp(4) + Math.exp(5) + Math.exp(6)))
-    ]);
-    const result = K.sparseCategoricalCrossentropy(target, x, true);
-    expectTensorsClose(result, expected);
-  });
-
-  it('from softmax', () => {
-    const x = tensor2d([[0.1, 0.2, 0.7], [0.2, 0.3, 0.5]], [2, 3]);
-    const target = tensor1d([0, 2]);
-    const expected = tensor1d([-1 * Math.log(0.1), -1 * Math.log(0.5)]);
-    const result = K.sparseCategoricalCrossentropy(target, x, false);
-    expectTensorsClose(result, expected);
-  });
-});
-
-describeMathCPUAndGPU('binaryCrossentropy', () => {
-  function _binaryCrossentropy(target: Tensor, output: Tensor): Tensor {
-    const targetComplement = K.scalarPlusArray(scalar(1), tfc.neg(target));
-    const outputComplement = K.scalarPlusArray(scalar(1), tfc.neg(output));
-    return tfc.neg(tfc.add(
-      tfc.mul(target, tfc.log(output)),
-      tfc.mul(targetComplement, tfc.log(outputComplement))));
-  }
-
-  it('from logits', () => {
-    const x = tensor2d([[1, 2], [3, 4]], [2, 2]);
-    const target = tensor2d([[0.25, 0.75], [0.1, 0.9]], [2, 2]);
-    const sigmoidX = tfc.sigmoid(x);
-    const expected = _binaryCrossentropy(target, sigmoidX);
-    const result = K.binaryCrossentropy(target, x, true);
-    expectTensorsClose(result, expected);
-  });
-
-  it('from sigmoid', () => {
-    const x = tensor2d([[0.3, 0.7], [0.4, 0.6]], [2, 2]);
-    const target = tensor2d([[0.25, 0.75], [0.1, 0.9]], [2, 2]);
-    const expected = _binaryCrossentropy(target, x);
-    const result = K.binaryCrossentropy(target, x, false);
-    expectTensorsClose(result, expected);
-  });
-});
-
-describeMathCPUAndGPU('sigmoidCrossEntropyWithLogits', () => {
-  it('outputs sigmoid cross-entropy', () => {
-    const x = tensor2d([[1, 2], [3, 4]], [2, 2]);
-    const target = tensor2d([[0.25, 0.75], [0.1, 0.9]], [2, 2]);
-    const targetComplement = K.scalarPlusArray(scalar(1), tfc.neg(target));
-    const sigmoidX = tfc.sigmoid(x);
-    const sigmoidXComplement = K.scalarPlusArray(scalar(1), tfc.neg(sigmoidX));
-    const expected = tfc.add(
-      tfc.mul(target, tfc.neg(tfc.log(sigmoidX))),
-      tfc.mul(targetComplement, tfc.neg(tfc.log(sigmoidXComplement))));
-    const result = K.sigmoidCrossEntropyWithLogits(target, x);
-    expectTensorsClose(result, expected);
-  });
-});
-
 describeMathCPUAndGPU('Sigmoid', () => {
   it('2D', () => {
     const xValues = [-5, -2, 0, 1, 2, 5];
